manager/requirements.txt
websocket-client 
construct
requests